test: Rename files according to convention
