+++ channel/test/types.go
@@ -16,6 +16,7 @@ package test
 
 import (
 	"math/rand"
+	"perun.network/go-perun/channel"
 	"perun.network/perun-cardano-backend/channel/types"
 )
 
@@ -30,4 +31,10 @@ func MakeRandomChannelState(rng *rand.Rand) types.ChannelState {
 		Version:  version,
 		Final:    final,
 	}
+}
+
+func MakeRandomChannelID(rng *rand.Rand) types.ID {
+	id := channel.ID{}
+	rng.Read(id[:])
+	return id
 }
